+++ lib/src/card_view.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/widgets.dart';
 
 import 'payment_html_view.dart';
+import 'square/error.dart';
 import 'square/payment_card.dart';
 
 /// Renders the Card or Gift Card form.
@@ -13,5 +14,8 @@ class CardView extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) => PaymentHtmlView(
-      onElementAttached: card.attach, onElementDetached: card.detach);
+      onElementAttached: (element) =>
+          card.attach(element).ignorePaymentMethodAlreadyDestroyedError(),
+      onElementDetached: () =>
+          card.detach().ignorePaymentMethodAlreadyDestroyedError(false));
 }