+++ lib/src/google_pay_view.dart
@@ -4,6 +4,7 @@ import 'package:flutter/widgets.dart';
 import 'package:square_web_payments/src/square/google_pay_button_options.dart';
 
 import 'payment_html_view.dart';
+import 'square/error.dart';
 import 'square/google_pay.dart';
 
 /// Renders the Google Pay button.
@@ -27,8 +28,11 @@ class GooglePayView extends StatelessWidget {
   @override
   Widget build(BuildContext context) => PaymentHtmlView(
       onElementAttached: (element) {
-        googlePay.attach(element, googlePayButtonOptions);
+        googlePay
+            .attach(element, googlePayButtonOptions)
+            .ignorePaymentMethodAlreadyDestroyedError();
         element.addEventListener('click', onPressed.toJS);
       },
-      onElementDetached: googlePay.detach);
+      onElementDetached: () =>
+          googlePay.detach().ignorePaymentMethodAlreadyDestroyedError(false));
 }