+++ lib/src/square/apple_pay.dart
@@ -1,8 +1,6 @@
-import 'dart:js_interop';
-
 import 'error.dart';
-import 'token_result.dart';
 import 'payment_method.dart';
+import 'token_result.dart';
 
 /// https://developer.squareup.com/reference/sdks/web/payments/objects/ApplePay
 class ApplePay implements PaymentMethod {
@@ -16,8 +14,7 @@ class ApplePay implements PaymentMethod {
 
 extension type JSApplePay._(JSPaymentMethod _) implements JSPaymentMethod {
   ApplePay get toDart => ApplePay(
-      destroy: () => destroy().toDart,
-      tokenize: () => tokenize().toDart.then(
-          (tokenResult) => tokenResult.toDart,
-          onError: (error) => throw (error as JSError).toDart));
+      destroy: () => tryCatchToDart(() => destroy()),
+      tokenize: () => tryCatchToDart(() => tokenize())
+          .then((tokenResult) => tokenResult.toDart));
 }