+++ lib/src/square/payment_card.dart
@@ -3,8 +3,8 @@ import 'dart:js_interop';
 import 'package:web/web.dart';
 
 import 'error.dart';
-import 'token_result.dart';
 import 'payment_method.dart';
+import 'token_result.dart';
 
 /// https://developer.squareup.com/reference/sdks/web/payments/objects/Card
 /// https://developer.squareup.com/reference/sdks/web/payments/objects/GiftCard
@@ -27,10 +27,10 @@ extension type JSPaymentCard._(JSPaymentMethod _) implements JSPaymentMethod {
   external JSPromise attach(HTMLDivElement element);
   external JSPromise<JSBoolean> detach();
   PaymentCard get toDart => PaymentCard(
-      attach: (element) => attach(element).toDart,
-      destroy: () => destroy().toDart,
-      detach: () => detach().toDart.then((result) => result.toDart),
-      tokenize: () => tokenize().toDart.then(
-          (tokenResult) => tokenResult.toDart,
-          onError: (error) => throw (error as JSError).toDart));
+      attach: (element) => tryCatchToDart(() => attach(element)),
+      destroy: () => tryCatchToDart(() => destroy()),
+      detach: () =>
+          tryCatchToDart(() => detach()).then((result) => result.toDart),
+      tokenize: () => tryCatchToDart(() => tokenize())
+          .then((tokenResult) => tokenResult.toDart));
 }