+++ CHANGELOG.md
@@ -47,3 +47,5 @@
 * Updating GitHub repository and pages URLs
 
 ## 0.0.12
+
+* Cleaning up error handling